tests: fix all flake8 issues
