+++ src/tests/intg/conftest.py
@@ -1,0 +1,1 @@
+from files_ops import passwd_ops_setup, group_ops_setup  # noqa
