+++ src/tests/multihost/ipa/conftest.py
@@ -7,11 +7,15 @@ import os
 import posixpath
 import pexpect
 import pytest
-from sssd.testlib.common.qe_class import session_multihost
 from sssd.testlib.common.utils import sssdTools
 from sssd.testlib.ipa.utils import ipaTools
 from sssd.testlib.common.utils import ADOperations
 from sssd.testlib.common.paths import SSSD_DEFAULT_CONF
+
+
+pytest_plugins = (
+    'sssd.testlib.common.fixtures',
+)
 
 
 def pytest_configure():
